refactor: 기존 테스트 이름을 변경하고 새 테스트 생성
